Consolidate PrintTask and ExportTask.
